Crate quantum

Advanced Rust quantum computer simulator.

See the code here.

Modules

algorithms
Examples of quantum algorithms.

complex
Complex number library code (public for pedagogical reasons).

computer
Main consumer module allowing easy control of whole quantum computer.

gate
Gate library code (public for pedagogical reasons).

gates
Implementations of quantum gates, intended for consumer use.

ket
Ket library code (public for pedagogical reasons).

macros
All library macros.

matrix
Matrix library code (public for pedagogical reasons).

other
(public for pedagogical reasons).

registers
Classical and quantum register library code (public for pedagogical reasons).

Macros

abs_square!
Compute a complex number's absolute value, i.e. |x + iy|2.

c!
Convenience macro for complex number construction.

m!
Convenience macro for matrix constuction.

m_real!
Convenience macro for constucting a matrix of reals.

square!
Square a numeric value efficiently by multiplying it with itself.